Refactor data menu and navigation.
